pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ragu"
version = "1.0.0"
description = "GraphRAG system"
requires-python = ">=3.10"
dependencies = [
    "aiolimiter>=1.2.1",
    "graspologic>=3.4.4",
    "instructor>=1.11.3",
    "loguru>=0.7.3",
    "nano-vectordb>=0.0.4.3",
    "networkx>=3.4.2",
    "numpy>=1.26.4",
    "openai>=1.109.1",
    "pandas>=2.3.3",
    "pandas-stubs==2.3.2.250926",
    "pydantic>=2.11.10",
    "pydantic-settings>=2.11.0",
    "pytest>=8.4.2",
    "razdel>=0.5.0",
    "tenacity>=9.1.2",
    "tiktoken>=0.11.0",
    "tqdm>=4.67.1",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
local = [
    "sentence-transformers>=5.1.1",
    "transformers>=4.57.0",
]

[tool.setuptools.packages.find]
include = ["ragu*"]



